Move getSiteComment and getSiteComments.
