+++ client/my-sites/comment/comment-delete-warning.jsx
@@ -17,7 +17,7 @@ import {
 	withAnalytics,
 } from 'state/analytics/actions';
 import { deleteComment } from 'state/comments/actions';
-import getSiteComment from 'state/selectors/get-site-comment';
+import { getSiteComment } from 'state/comments/selectors';
 import Notice from 'components/notice';
 import NoticeAction from 'components/notice/notice-action';
