+++ client/my-sites/comment/comment-permalink.jsx
@@ -13,7 +13,7 @@ import { get, isUndefined } from 'lodash';
 import { Card } from '@automattic/components';
 import SectionHeader from 'components/section-header';
 import ExternalLink from 'components/external-link';
-import getSiteComment from 'state/selectors/get-site-comment';
+import { getSiteComment } from 'state/comments/selectors';
 
 const CommentPermalink = ( { isLoading, permaLink, translate } ) =>
 	! isLoading && (
