+++ client/my-sites/comments/comment/comment-content.jsx
@@ -18,7 +18,7 @@ import Emojify from 'components/emojify';
 import QueryComment from 'components/data/query-comment';
 import { stripHTML, decodeEntities } from 'lib/formatting';
 import getParentComment from 'state/selectors/get-parent-comment';
-import getSiteComment from 'state/selectors/get-site-comment';
+import { getSiteComment } from 'state/comments/selectors';
 import { isJetpackSite } from 'state/sites/selectors';
 import { getSelectedSiteId, getSelectedSiteSlug } from 'state/ui/selectors';
